vision/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pillow==10.4.0
numpy==1.26.4
